fix(build): resolve critical compilation errors blocking CI pipeline
- Consolidated duplicate type declarations in api/v1 packages
- Added missing type implementations in security and audit packages
- Fixed 100+ syntax errors with escaped quotes in multiple files
- Resolved import conflicts and missing imports
- Fixed type compatibility issues with Kubernetes APIs
- Renamed conflicting types to unique identifiers

This addresses the core compilation failures preventing CI from running.
Additional package-specific issues remain but are non-critical.

Affected packages: api/v1, pkg/audit, pkg/security, pkg/errors
Modified: 24 files, +8252/-1487 lines

+++ pkg/audit/middleware.go
@@ -315,7 +315,7 @@ func (m *HTTPAuditMiddleware) captureRequestBody(r *http.Request) (string, error
 		bodyString := string(bodyBytes) + "[TRUNCATED]"
 		
 		// Replace body so subsequent handlers can read it
-		r.Body = io.NopCloser(strings.NewReader(string(bodyBytes[:m.config.MaxBodySize-len("[TRUNCATED]")])))
+		r.Body = io.NopCloser(strings.NewReader(string(bodyBytes[:m.config.MaxBodySize-int64(len("[TRUNCATED]"))])))
 		return bodyString, nil
 	}
